[Sylius/Behat] Extracted SymfonyPageObjectExtension from PageObjectExtension
