+++ src/Sylius/Behat/Page/Cart/CartSummaryPage.php
@@ -11,7 +11,7 @@
 
 namespace Sylius\Behat\Page\Cart;
 
-use Sylius\Behat\PageObjectExtension\Page\SymfonyPage;
+use Sylius\Behat\SymfonyPageObjectExtension\Page\SymfonyPage;
 
 /**
  * @author Mateusz Zalewski <mateusz.zalewski@lakion.com>
