+++ src/Sylius/Behat/Page/Checkout/CheckoutShippingStep.php
@@ -11,7 +11,7 @@
 
 namespace Sylius\Behat\Page\Checkout;
 
-use Sylius\Behat\PageObjectExtension\Page\SymfonyPage;
+use Sylius\Behat\SymfonyPageObjectExtension\Page\SymfonyPage;
 
 /**
  * @author Arkadiusz Krakowiak <arkadiusz.krakowiak@lakion.com>
