+++ src/Sylius/Behat/SymfonyPageObjectExtension/Page/SymfonyPage.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Sylius\Behat\PageObjectExtension\Page;
+namespace Sylius\Behat\SymfonyPageObjectExtension\Page;
 
 use Behat\Mink\Session;
 use SensioLabs\Behat\PageObjectExtension\PageObject\Exception\ElementNotFoundException;
