+++ src/Sylius/Behat/spec/Page/Cart/CartSummaryPageSpec.php
@@ -14,7 +14,7 @@ namespace spec\Sylius\Behat\Page\Cart;
 use Behat\Mink\Session;
 use PhpSpec\ObjectBehavior;
 use SensioLabs\Behat\PageObjectExtension\PageObject\Factory;
-use Sylius\Behat\PageObjectExtension\Page\SymfonyPage;
+use Sylius\Behat\SymfonyPageObjectExtension\Page\SymfonyPage;
 use Symfony\Component\Routing\RouterInterface;
 
 /**
